Refactor: extract immediate input handling into separate manager class and interface
